Propagate blocking function information through function literal calls.
When a function literal is used that calls another function and
that other function has yet to be analyzed, the call site for
the function literal is not retained, so it's not processed later,
so the function that was calling the function literal is not made
blocking itself, when it should be.

This change makes blocking status propagate through any number of
intermediate literal function calls.

This is a rebase of https://github.com/gopherjs/gopherjs/pull/956.

Fixes https://github.com/gopherjs/gopherjs/issues/955.

Co-authored-by: Frank Rehwinkel <frankrehwinkel@gmail.com>